refactor(voice-chat): replace two-panel layout with unified conversation view
Extract VoiceUserBubble, VoiceAssistantBubble, SlowBrainIndicator into
shared voice-chat-bubbles.tsx, imported by both MC panel and standalone page.
Remove push-to-talk functionality. Simplify setupWebRTC$ to always use
semantic_vad. Replace grid-cols-2 split with single scrollable conversation
column using chat bubbles with inline slow-brain indicators. Simplify footer
to Mute/Unmute + End Session + Retry.

Closes #9587
