+++ turbo/apps/platform/src/signals/voice-chat/voice-chat-session.ts
@@ -221,7 +221,6 @@ const internalPrompt$ = state<string | null>(null);
 const internalPrepStartTime$ = state<number | null>(null);
 const internalPrepElapsedMs$ = state(0);
 const internalReconnectAttempt$ = state(0);
-const internalInputMode$ = state<"hands-free" | "push-to-talk">("hands-free");
 const internalModel$ = state<RealtimeModel>("gpt-realtime-mini");
 
 const internalParentSignal$ = state<AbortSignal | null>(null);
@@ -242,6 +241,13 @@ export const vcTranscript$ = computed((get) => {
 export const vcEvents$ = computed((get) => {
   return get(internalEvents$);
 });
+
+export const vcSlowBrainEvents$ = computed((get) => {
+  return get(internalEvents$).filter((e) => {
+    return e.source === "slow-brain";
+  });
+});
+
 export const vcMuted$ = computed((get) => {
   return get(internalMuted$);
 });
@@ -263,9 +269,6 @@ export const vcPrepElapsedMs$ = computed((get) => {
 });
 export const vcReconnectAttempt$ = computed((get) => {
   return get(internalReconnectAttempt$);
-});
-export const vcInputMode$ = computed((get) => {
-  return get(internalInputMode$);
 });
 export const vcMeetingPromptInput$ = computed((get) => {
   return get(meetingPromptInput$);
@@ -527,10 +530,6 @@ const setupWebRTC$ = command(
     set(internalDc$, dc);
 
     dc.addEventListener("open", () => {
-      const inputMode = get(internalInputMode$);
-      const turnDetection =
-        inputMode === "hands-free" ? HANDS_FREE_VAD_CONFIG : null;
-
       dc.send(
         JSON.stringify({
           type: "session.update",
@@ -539,7 +538,7 @@ const setupWebRTC$ = command(
             instructions: FAST_BRAIN_INSTRUCTIONS,
             input_audio_transcription: { model: "gpt-4o-mini-transcribe" },
             input_audio_noise_reduction: { type: "far_field" },
-            turn_detection: turnDetection,
+            turn_detection: HANDS_FREE_VAD_CONFIG,
             tools: SESSION_TOOLS,
           },
         }),
@@ -1446,7 +1445,6 @@ export const endVoiceChat$ = command(({ get, set }) => {
   set(internalPrepStartTime$, null);
   set(internalPrepElapsedMs$, 0);
   set(internalReconnectAttempt$, 0);
-  set(internalInputMode$, "hands-free");
   set(internalModel$, "gpt-realtime-mini");
   set(internalParentSignal$, null);
   set(internalStatus$, "idle");
@@ -1474,74 +1472,4 @@ export const toggleVoiceChatMute$ = command(({ get, set }) => {
   const wasMuted = get(internalMuted$);
   track.enabled = wasMuted;
   set(internalMuted$, !wasMuted);
-});
-
-export const switchInputMode$ = command(
-  ({ get, set }, mode: "hands-free" | "push-to-talk") => {
-    const dc = get(internalDc$);
-    if (!dc || dc.readyState !== "open") {
-      return;
-    }
-
-    set(internalInputMode$, mode);
-
-    const turnDetection = mode === "hands-free" ? HANDS_FREE_VAD_CONFIG : null;
-
-    dc.send(
-      JSON.stringify({
-        type: "session.update",
-        session: { turn_detection: turnDetection },
-      }),
-    );
-
-    // PTT starts muted, hands-free starts unmuted
-    const stream = get(internalStream$);
-    if (!stream) {
-      return;
-    }
-    const track = stream.getAudioTracks()[0];
-    if (!track) {
-      return;
-    }
-
-    if (mode === "push-to-talk") {
-      track.enabled = false;
-      set(internalMuted$, true);
-    } else {
-      track.enabled = true;
-      set(internalMuted$, false);
-    }
-  },
-);
-
-export const startPTT$ = command(({ get, set }) => {
-  const stream = get(internalStream$);
-  if (!stream) {
-    return;
-  }
-  const track = stream.getAudioTracks()[0];
-  if (!track) {
-    return;
-  }
-  track.enabled = true;
-  set(internalMuted$, false);
-});
-
-export const stopPTT$ = command(({ get, set }) => {
-  const stream = get(internalStream$);
-  if (!stream) {
-    return;
-  }
-  const track = stream.getAudioTracks()[0];
-  if (!track) {
-    return;
-  }
-  track.enabled = false;
-  set(internalMuted$, true);
-
-  const dc = get(internalDc$);
-  if (dc && dc.readyState === "open") {
-    dc.send(JSON.stringify({ type: "input_audio_buffer.commit" }));
-    dc.send(JSON.stringify({ type: "response.create" }));
-  }
 });
